pyproject.toml
[tool.poetry]
name = "tele-track"
version = "0.1.0"
description = ""
authors = ["Andrii Oriekhov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.36.0"
pytelegrambotapi = "^4.22.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
